refactor(wallets): extract the Rango-to-CAIP chain mapping from wallets-shared
Move the chain-vocabulary translation into `@rango-dev/wallets-blockchains` and
the CAIP-2 registry it uses into `@hub3js/caip`, so provider metadata and
namespace code depend only on the CAIP half and signers on the Rango half.
`Networks` moves with the mapping and wallets-shared re-exports it;
`getSupportedChainsFromNamespace` and `getSupportedChainsFromProvider` stay.

Declare the wallets-shared dependency provider-noir-wallet was importing without
listing, and switch walletconnect-2's UTXO signer to the shared
`isBitcoinBlockchain` guard.
