+++ wallets/blockchains/package.json
@@ -1,14 +1,17 @@
 {
-  "name": "@rango-dev/provider-xverse",
-  "version": "0.8.0",
+  "name": "@rango-dev/wallets-blockchains",
+  "version": "0.1.0",
   "license": "MIT",
   "type": "module",
   "source": "./src/mod.ts",
   "main": "./dist/mod.js",
-  "exports": {
-    ".": "./dist/mod.js"
-  },
   "typings": "dist/mod.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/mod.d.ts",
+      "default": "./dist/mod.js"
+    }
+  },
   "files": [
     "dist",
     "src"
@@ -18,15 +21,13 @@
     "ts-check": "tsc --declaration --emitDeclarationOnly -p ./tsconfig.json",
     "clean": "rimraf dist",
     "format": "prettier --write '{.,src}/**/*.{ts,tsx}'",
-    "lint": "eslint \"**/*.{ts,tsx}\""
+    "lint": "eslint \"**/*.{ts,tsx}\"",
+    "test": "vitest",
+    "test:watch": "vitest --watch",
+    "test:coverage": "vitest run --coverage"
   },
   "dependencies": {
-    "@hub3js/core": "^0.61.1",
-    "@hub3js/namespaces": "^0.2.5",
-    "@hub3js/std": "0.0.0-experimental-53df86ae-20260819",
-    "@rango-dev/wallets-blockchains": "^0.1.0",
-    "@rango-dev/wallets-core": "^0.61.0",
-    "@rango-dev/wallets-shared": "^0.62.0",
+    "@hub3js/caip": "0.0.0-experimental-123b0f23-20260819",
     "rango-types": "^0.5.0"
   },
   "publishConfig": {
