+++ wallets/blockchains/tsconfig.json
@@ -1,0 +1,1 @@
+{ "extends": "./tsconfig.build.json", "include": ["src", "tests"] }
