+++ wallets/provider-bitget/package.json
@@ -28,6 +28,7 @@
     "@rango-dev/common-core": "^0.1.0",
     "@rango-dev/signer-evm": "^0.45.0",
     "@rango-dev/signer-tron": "^0.42.0",
+    "@rango-dev/wallets-blockchains": "^0.1.0",
     "@rango-dev/wallets-core": "^0.61.0",
     "@rango-dev/wallets-shared": "^0.62.0",
     "rango-types": "^0.5.0"
