+++ wallets/provider-ctrl/package.json
@@ -21,6 +21,7 @@
     "lint": "eslint \"**/*.{ts,tsx}\""
   },
   "dependencies": {
+    "@hub3js/caip": "0.0.0-experimental-123b0f23-20260819",
     "@hub3js/core": "^0.61.1",
     "@hub3js/evm": "^0.3.3",
     "@hub3js/namespaces": "^0.2.5",
@@ -28,6 +29,7 @@
     "@hub3js/std": "0.0.0-experimental-53df86ae-20260819",
     "@rango-dev/signer-evm": "^0.45.0",
     "@rango-dev/signer-solana": "^0.49.0",
+    "@rango-dev/wallets-blockchains": "^0.1.0",
     "@rango-dev/wallets-core": "^0.61.0",
     "@rango-dev/wallets-shared": "^0.62.0",
     "bs58": "^5.0.0",
