+++ wallets/provider-phantom/package.json
@@ -33,6 +33,7 @@
     "@rango-dev/common-core": "^0.1.0",
     "@rango-dev/signer-solana": "^0.49.0",
     "@rango-dev/signer-sui": "^0.11.0",
+    "@rango-dev/wallets-blockchains": "^0.1.0",
     "@rango-dev/wallets-core": "^0.61.0",
     "@rango-dev/wallets-shared": "^0.62.0",
     "bitcoinjs-lib": "^6.1.7",
