+++ wallets/provider-unisat/package.json
@@ -25,6 +25,7 @@
     "@hub3js/namespaces": "^0.2.5",
     "@hub3js/std": "0.0.0-experimental-53df86ae-20260819",
     "@rango-dev/common-core": "^0.1.0",
+    "@rango-dev/wallets-blockchains": "^0.1.0",
     "@rango-dev/wallets-core": "^0.61.0",
     "@rango-dev/wallets-shared": "^0.62.0",
     "bitcoinjs-lib": "^6.1.7",
