+++ wallets/provider-vultisig/package.json
@@ -24,6 +24,7 @@
     "@hub3js/core": "^0.61.1",
     "@hub3js/namespaces": "^0.2.5",
     "@hub3js/std": "0.0.0-experimental-53df86ae-20260819",
+    "@rango-dev/wallets-blockchains": "^0.1.0",
     "@rango-dev/wallets-core": "^0.61.0",
     "@rango-dev/wallets-shared": "^0.62.0",
     "rango-types": "^0.5.0"
