+++ wallets/provider-walletconnect-2/package.json
@@ -21,19 +21,20 @@
     "lint": "eslint \"**/*.{ts,tsx}\""
   },
   "dependencies": {
-    "@solana/web3.js": "^1.91.4",
-    "@walletconnect/encoding": "^1.0.2",
+    "@bitcoinerlab/secp256k1": "^1.2.0",
     "@hub3js/core": "^0.61.1",
     "@hub3js/evm": "^0.3.3",
     "@hub3js/std": "0.0.0-experimental-53df86ae-20260819",
-    "@bitcoinerlab/secp256k1": "^1.2.0",
-    "@reown/appkit": "1.8.20",
     "@rango-dev/common-core": "^0.1.0",
     "@rango-dev/logging-core": "^0.14.0",
     "@rango-dev/signer-evm": "^0.45.0",
     "@rango-dev/signer-solana": "^0.49.0",
+    "@rango-dev/wallets-blockchains": "^0.1.0",
     "@rango-dev/wallets-core": "^0.60.1-next.3",
     "@rango-dev/wallets-shared": "^0.62.0",
+    "@reown/appkit": "1.8.20",
+    "@solana/web3.js": "^1.91.4",
+    "@walletconnect/encoding": "^1.0.2",
     "@walletconnect/sign-client": "2.23.7",
     "@walletconnect/universal-provider": "2.23.7",
     "@walletconnect/utils": "2.23.7",
